+++ supervisor/addons/__init__.py
@@ -280,7 +280,15 @@ class AddonManager(CoreSysAttributes):
         # Check if available, Maybe something have changed
         if not store.available:
             _LOGGER.error("Add-on %s not supported on that platform", slug)
-            raise AddonsNotSupportedError()
+
+            if not store.available_arch:
+                raise AddonsNotSupportedError(
+                    "Architecture is not supported by the add-on"
+                )
+            if not store.available_homeassistant:
+                raise AddonsNotSupportedError(
+                    f"Home Assistant version is not supported by the add-on, you need at least version {store.homeassistant_version}"
+                )
 
         # Update instance
         last_state: AddonState = addon.state
