+++ supervisor/addons/model.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 from pathlib import Path
 from typing import Any, Awaitable, Dict, List, Optional
 
-from awesomeversion import AwesomeVersion, AwesomeVersionException
+from awesomeversion import AwesomeVersion
 import voluptuous as vol
 
 from ..const import (
@@ -101,21 +101,6 @@ class AddonModel(CoreSysAttributes, ABC):
     @abstractmethod
     def is_detached(self) -> bool:
         """Return True if add-on is detached."""
-
-    @property
-    def available_arch(self) -> bool:
-        """Return True if the running arch is supported by the add-on."""
-        return self._available_arch(self.data)
-
-    @property
-    def available_homeassistant(self) -> bool:
-        """Return True if the running version of Home Assistant is supported by the add-on."""
-        return self._available_homeassistant(self.data)
-
-    @property
-    def available(self) -> bool:
-        """Return True if this add-on is available on this platform."""
-        return self._available(self.data)
 
     @property
     def options(self) -> Dict[str, Any]:
@@ -559,33 +544,6 @@ class AddonModel(CoreSysAttributes, ABC):
         if not isinstance(other, AddonModel):
             return False
         return self.slug == other.slug
-
-    def _available_arch(self, config) -> bool:
-        """Return True if the running arch is supported by the add-on."""
-        # Architecture
-        if not self.sys_arch.is_supported(config[ATTR_ARCH]):
-            return False
-
-        # Machine / Hardware
-        machine = config.get(ATTR_MACHINE)
-        if machine and f"!{self.sys_machine}" in machine:
-            return False
-        elif machine and self.sys_machine not in machine:
-            return False
-
-        return True
-
-    def _available_homeassistant(self, config) -> bool:
-        """Return True if the running version of Home Assistant is supported by the add-on."""
-        version: Optional[AwesomeVersion] = config.get(ATTR_HOMEASSISTANT)
-        try:
-            return self.sys_homeassistant.version >= version
-        except (AwesomeVersionException, TypeError):
-            return True
-
-    def _available(self, config) -> bool:
-        """Return True if this add-on is available on this platform."""
-        return self._available_arch(config) and self._available_homeassistant(config)
 
     def _image(self, config) -> str:
         """Generate image name from data."""
